+++ js/filter.js
@@ -1,17 +1,17 @@
-import {addClassName, disableField, enableField, removeClassName} from './util.js';
+import {disableField, enableField} from './util.js';
 
 const filter = document.querySelector('.map__filters');
 const filterFieldsets = filter.querySelectorAll('fieldset');
 
 function deactivateFilter () {
-  addClassName(filter, 'map__filters--disabled');
+  filter.classList.add('map__filters--disabled');
   filterFieldsets.forEach( (element) => {
     disableField(element);
   } );
 }
 
 function activateFilter () {
-  removeClassName(filter, 'map__filters--disabled');
+  filter.classList.remove('map__filters--disabled');
   filterFieldsets.forEach( (element) => {
     enableField(element);
   } );
